Implements account in WASM
