+++ .gitignore
@@ -2,3 +2,4 @@
 **/target
 **.DS_Store
 storage*/
+wasm/Cargo.lock
