+++ account/benches/account.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the Aleo library. If not, see <https://www.gnu.org/licenses/>.
 
-use aleo::account::{Address, PrivateKey};
+use aleo_account::{Address, PrivateKey};
 
 #[macro_use]
 extern crate bencher;
