+++ src/errors.rs
@@ -15,60 +15,6 @@
 // along with the Aleo library. If not, see <https://www.gnu.org/licenses/>.
 
 #[derive(Debug, Error)]
-pub enum PrivateKeyError {
-    #[error("{}: {}", _0, _1)]
-    Crate(&'static str, String),
-}
-
-impl From<snarkvm_errors::objects::account::AccountError> for PrivateKeyError {
-    fn from(error: snarkvm_errors::objects::account::AccountError) -> Self {
-        PrivateKeyError::Crate("snarkvm_errors::objects::account", format!("{:?}", error))
-    }
-}
-
-impl From<std::io::Error> for PrivateKeyError {
-    fn from(error: std::io::Error) -> Self {
-        PrivateKeyError::Crate("std::io", format!("{:?}", error))
-    }
-}
-
-#[derive(Debug, Error)]
-pub enum ViewKeyError {
-    #[error("{}: {}", _0, _1)]
-    Crate(&'static str, String),
-}
-
-impl From<snarkvm_errors::objects::account::AccountError> for ViewKeyError {
-    fn from(error: snarkvm_errors::objects::account::AccountError) -> Self {
-        ViewKeyError::Crate("snarkvm_errors::objects::account", format!("{:?}", error))
-    }
-}
-
-impl From<std::io::Error> for ViewKeyError {
-    fn from(error: std::io::Error) -> Self {
-        ViewKeyError::Crate("std::io", format!("{:?}", error))
-    }
-}
-
-#[derive(Debug, Error)]
-pub enum AddressError {
-    #[error("{}: {}", _0, _1)]
-    Crate(&'static str, String),
-}
-
-impl From<snarkvm_errors::objects::account::AccountError> for AddressError {
-    fn from(error: snarkvm_errors::objects::account::AccountError) -> Self {
-        AddressError::Crate("snarkvm_errors::objects::account", format!("{:?}", error))
-    }
-}
-
-impl From<std::io::Error> for AddressError {
-    fn from(error: std::io::Error) -> Self {
-        AddressError::Crate("std::io", format!("{:?}", error))
-    }
-}
-
-#[derive(Debug, Error)]
 pub enum UpdaterError {
     #[error("{}: {}", _0, _1)]
     Crate(&'static str, String),
