+++ src/lib.rs
@@ -17,7 +17,6 @@
 #[macro_use]
 extern crate thiserror;
 
-pub mod account;
 pub mod cli;
 pub mod commands;
 pub mod errors;
