+++ wasm/Cargo.toml
@@ -15,7 +15,9 @@ edition = "2018"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-aleo = { path = ".." }
+aleo-account = { path = "../account" }
+
+rand = { version = "0.7", default-features = false, features = ["wasm-bindgen"] }
 wasm-bindgen = { version = "0.2.70" }
 
 [dev-dependencies]
